Move `BaseOperatorLink` into the separate module
