+++ STATIC_CODE_CHECKS.rst
@@ -154,9 +154,6 @@ require Breeze Docker image to be built locally.
 +-----------------------------------------------------------+--------------------------------------------------------------+---------+
 | check-base-operator-partial-arguments                     | Check BaseOperator and partial() arguments                   |         |
 +-----------------------------------------------------------+--------------------------------------------------------------+---------+
-| check-base-operator-usage                                 | * Check BaseOperator[Link] core imports                      |         |
-|                                                           | * Check BaseOperator[Link] other imports                     |         |
-+-----------------------------------------------------------+--------------------------------------------------------------+---------+
 | check-boring-cyborg-configuration                         | Checks for Boring Cyborg configuration consistency           |         |
 +-----------------------------------------------------------+--------------------------------------------------------------+---------+
 | check-breeze-top-dependencies-limited                     | Breeze should have small number of top-level dependencies    |         |
